nasa_exoplanet_detector/requirements.txt
Django==4.2.13
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
djangorestframework==3.15.2
requests==2.32.3
